third_party/WebKit/LayoutTests/fast/js/Promise-native-then-expected.txt

Test Promise native then.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Got a new promise: [object Promise]
PASS result is 11
PASS result is 12
PASS DONE
PASS successfullyParsed is true

TEST COMPLETE
